+++ src/game/level-settings/hardSettings.ts
@@ -1,13 +1,13 @@
-import { LevelSettings } from "../level-creation/generateRandomLevel";
+import { SettingsProducer } from "../level-creation/generateRandomLevel";
 
 import { getNormalSettings } from "./normalSettings";
 
-export const getHardSettings = (difficulty: number): LevelSettings => ({
+export const getHardSettings: SettingsProducer = (difficulty) => ({
   ...getNormalSettings(difficulty),
   hideBlockTypes: "all",
 });
 
-export const getHard2Settings = (difficulty: number): LevelSettings => ({
+export const getHard2Settings: SettingsProducer = (difficulty) => ({
   ...getNormalSettings(difficulty),
   hideBlockTypes: "checker",
 });