+++ CHANGELOG.md
@@ -5,15 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Version 1.6.2 - UNRELEASED
+## Version 1.7.0 - UNRELEASED
 
 ### Added
 
 - Labels to options in 'zen mode' screen
+- New special level type, with steps buffer
 
 ### Fixed
 
 - Always ask for confirmation before import
+
+### Changed
+
+- Updated the difficulty of some of the special templates
+
+### Removed
+
+- The beta label on the data transfer
 
 ## Version 1.6.1 - 2024-10-08